# Added numa options to allow finer grained control as well as plumbing for a new mirror mode that will require numa.h