--- a/Makefile
+++ b/Makefile
@@ -265,8 +265,8 @@ ifndef RISCV
 
 ifeq ($(UNAME_M),$(filter $(UNAME_M),x86_64 i686 amd64))
 	# Use all CPU extensions that are available:
-	MK_CFLAGS     += -march=native -mtune=native
-	HOST_CXXFLAGS += -march=native -mtune=native
+	MK_CFLAGS     += -march=znver4 -mtune=znver4
+	HOST_CXXFLAGS += -march=znver4 -mtune=znver4
 
 	# Usage AVX-only
 	#MK_CFLAGS   += -mfma -mf16c -mavx