--- a/common/common.cpp
+++ b/common/common.cpp
@@ -674,13 +674,12 @@ bool gpt_params_parse_ex(int argc, char ** argv, gpt_params & params) {
             if (++i >= argc) {
                 invalid_param = true;
                 break;
-            } else {
-               std::string value(argv[i]);
-               /**/ if (value == "distribute" || value == "" ) { params.numa = GGML_NUMA_STRATEGY_DISTRIBUTE; }
-               else if (value == "isolate") { params.numa = GGML_NUMA_STRATEGY_ISOLATE; }
-               else if (value == "numactl") { params.numa = GGML_NUMA_STRATEGY_NUMACTL; }
-               else { invalid_param = true; break; }
             }
+            std::string value(argv[i]);
+            /**/ if (value == "distribute" || value == "" ) { params.numa = GGML_NUMA_STRATEGY_DISTRIBUTE; }
+            else if (value == "isolate") { params.numa = GGML_NUMA_STRATEGY_ISOLATE; }
+            else if (value == "numactl") { params.numa = GGML_NUMA_STRATEGY_NUMACTL; }
+            else { invalid_param = true; break; }
         } else if (arg == "--verbose-prompt") {
             params.verbose_prompt = true;
         } else if (arg == "--no-display-prompt") {