… examples/batched.swift/Sources/main.swift
… examples/batched/batched.cpp
… examples/beam-search/beam-search.cpp
… examples/embedding/embedding.cpp
… examples/imatrix/imatrix.cpp
… examples/infill/infill.cpp
… examples/llama-bench/llama-bench.cpp
… examples/llama.android/app/src/main/cpp/llama-android.cpp
… examples/llama.swiftui/llama.cpp.swift/LibLlama.swift
… examples/llava/llava-cli.cpp
… examples/lookahead/lookahead.cpp
… examples/lookup/lookup.cpp
… examples/main/README.md
… examples/main/main.cpp
… examples/parallel/parallel.cpp
… examples/passkey/passkey.cpp
… examples/perplexity/perplexity.cpp
… examples/quantize/quantize.cpp
… examples/server/README.md
… examples/server/server.cpp
… examples/simple/simple.cpp
… examples/speculative/speculative.cpp
… examples/tokenize/tokenize.cpp
… ggml.h
… llama.cpp
… llama.h
… tests/test-autorelease.cpp
… tests/test-model-load-cancel.cpp
… tests/test-tokenizer-0-falcon.cpp
… tests/test-tokenizer-0-llama.cpp