Attempt number two. 

Removed sched.h from ggml.h, moved ggml_get_numa_affinity out of the public API and purely into ggml.c, removed trailing whitespace and fixed up a few inconsistent variables