@bmtwl 2024-02-07
> If mirror mode isn't implemented yet, the user should be shown a warning or error if they try to use it - "Mirror Mode Enabled" doesn't communicate that.

I figured that hiding it behind the #ifdef would be enough, but I can add a warning in for sure