@cebtenzzre 2024-02-07
> I figured that hiding it behind the #ifdef would be enough, but I can add a warning in for sure

If there is currently no use for #defining GGML_NUMA_MIRROR, then the code that depends on it shouldn't be committed yet.