@bmtwl 2024-02-08
I have fixed the errors in the last test. I also fixed a few related errors in the "examples" folder
All tests now pass in both make and cmake :

100% tests passed, 0 tests failed out of 22