@rankaiyx 2024-02-12
Test environment:
E5-2698Bv3 x 2 (Each one has 16 cores 32 threads)
DDR3 1866 32G x 8  (Bandwidth 100GB/s)

Bmtwl branch:
<details>
$ sudo sh -c "/usr/bin/echo 3 > /proc/sys/vm/drop_caches"

$ numactl -H
available: 2 nodes (0-1)
node 0 cpus: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47
node 0 size: 128788 MB
node 0 free: 128163 MB
node 1 cpus: 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 48 49 50 51 52 53 54 55 56 57 58 59 60 61 62 63
node 1 size: 129010 MB
node 1 free: 128514 MB
node distances:
node   0   1
  0:  10  21
  1:  21  10

$ ./main -m ../../text-generation-webui/models/miqu-1-70b.q4_k_m.gguf -s 12345 -n 100 -t 30 --numa isolate -p "I believe the meaning of life is"

 I believe the meaning of life is to be happy and content in what we do. For me, this means working at a job that is not only rewarding but also challenges me every day. I have been lucky enough to find such a job, and I am grateful for it every day.

I also believe that the meaning of life is to help others whenever possible. This can be done in many ways, big or small. Whether it's volunteering at a soup kitchen or simply holding the door open for
llama_print_timings:        load time =   21157.95 ms
llama_print_timings:      sample time =      60.81 ms /   100 runs   (    0.61 ms per token,  1644.47 tokens per second)
llama_print_timings: prompt eval time =    2149.83 ms /     8 tokens (  268.73 ms per token,     3.72 tokens per second)
llama_print_timings:        eval time =   80266.56 ms /    99 runs   (  810.77 ms per token,     1.23 tokens per second)
llama_print_timings:       total time =   82523.55 ms /   107 tokens
Log end

$ numactl -H
available: 2 nodes (0-1)
node 0 cpus: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47
node 0 size: 128788 MB
node 0 free: 88696 MB
node 1 cpus: 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 48 49 50 51 52 53 54 55 56 57 58 59 60 61 62 63
node 1 size: 129010 MB
node 1 free: 128513 MB
node distances:
node   0   1
  0:  10  21
  1:  21  10
(base)


$ sudo sh -c "/usr/bin/echo 3 > /proc/sys/vm/drop_caches"
$ ./main -m ../../text-generation-webui/models/miqu-1-70b.q4_k_m.gguf -s 12345 -n 100 -t 30 --numa interleave -p "I believe the meaning of life is"
 I believe the meaning of life is to be happy and content in what we do. For me, this means working at a job that is not only rewarding but also challenges me every day. I have been lucky enough to find such a job, and I am grateful for it every day.

I also believe that the meaning of life is to help others whenever possible. This can be done in many ways, big or small. Whether it's volunteering at a soup kitchen or simply holding the door open for
llama_print_timings:        load time =   24331.70 ms
llama_print_timings:      sample time =      64.47 ms /   100 runs   (    0.64 ms per token,  1551.06 tokens per second)
llama_print_timings: prompt eval time =    1543.93 ms /     8 tokens (  192.99 ms per token,     5.18 tokens per second)
llama_print_timings:        eval time =   63096.96 ms /    99 runs   (  637.34 ms per token,     1.57 tokens per second)
llama_print_timings:       total time =   64755.88 ms /   107 tokens
Log end

$ numactl -H
available: 2 nodes (0-1)
node 0 cpus: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47
node 0 size: 128788 MB
node 0 free: 108383 MB
node 1 cpus: 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 48 49 50 51 52 53 54 55 56 57 58 59 60 61 62 63
node 1 size: 129010 MB
node 1 free: 108821 MB
</details>


Mainline branch:
<details>
$ sudo sh -c "/usr/bin/echo 3 > /proc/sys/vm/drop_caches"
$ ./main -m ../../text-generation-webui/models/miqu-1-70b.q4_k_m.gguf -s 12345 -n 100 -t 30 --numa  -p "I believe the meaning of life is"
 I believe the meaning of life is to be happy and content in what we do. For me, this means working at a job that is not only rewarding but also challenges me every day. I have been lucky enough to find such a job, and I am grateful for it every day.

I also believe that the meaning of life is to help others whenever possible. This can be done in many ways, big or small. Whether it's volunteering at a soup kitchen or simply holding the door open for
llama_print_timings:        load time =   21422.24 ms
llama_print_timings:      sample time =      64.49 ms /   100 runs   (    0.64 ms per token,  1550.65 tokens per second)
llama_print_timings: prompt eval time =    1570.82 ms /     8 tokens (  196.35 ms per token,     5.09 tokens per second)
llama_print_timings:        eval time =   60025.33 ms /    99 runs   (  606.32 ms per token,     1.65 tokens per second)
llama_print_timings:       total time =   61708.69 ms /   107 tokens
Log end

$ numactl -H
available: 2 nodes (0-1)
node 0 cpus: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47
node 0 size: 128788 MB
node 0 free: 108386 MB
node 1 cpus: 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 48 49 50 51 52 53 54 55 56 57 58 59 60 61 62 63
node 1 size: 129010 MB
node 1 free: 108807 MB
</details>