@rankaiyx 2024-02-12
Mainline branch  with numactl:
<details>
$ numactl --interleave=0-1 ./main -m ../../text-generation-webui/models/miqu-1-70b.q4_k_m.gguf -s 12345 -n 100 -t 30  -p "I believe the meaning of life is"

 I believe the meaning of life is to be happy and content in what we do. For me, this means working at a job that is not only rewarding but also challenges me every day. I have been lucky enough to find such a job, and I am grateful for it every day.

I also believe that the meaning of life is to help others whenever possible. This can be done in many ways, big or small. Whether it's volunteering at a soup kitchen or simply holding the door open for
llama_print_timings:        load time =   25683.35 ms
llama_print_timings:      sample time =      68.15 ms /   100 runs   (    0.68 ms per token,  1467.29 tokens per second)
llama_print_timings: prompt eval time =    1726.15 ms /     8 tokens (  215.77 ms per token,     4.63 tokens per second)
llama_print_timings:        eval time =   60565.81 ms /    99 runs   (  611.78 ms per token,     1.63 tokens per second)
llama_print_timings:       total time =   62408.81 ms /   107 tokens
Log end
</details>

I wonder why the performance difference is so small now between the way the numa option is used in the mainline branch and the way numactl is used? Obviously there was quite a bit of difference before, and the numa option is about 1.3x faster than numactl.