@rankaiyx 2024-02-13
Run the following two inferences in parallel:
$ numactl -N0 -m0 ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 16 --numa numactl -p "I believe the meaning of life is" --no-mmap -b 65535 -c 4096 -np 4096 -ns 65535 -cb
$ numactl -N1 -m1 ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M_2.gguf -s 12346 -n 100 -t 16 --numa numactl -p "I believe the meaning of life is" --no-mmap -b 65535 -c 4096 -np 4096 -ns 65535 -cb

llama_print_timings:        load time =    4198.93 ms
llama_print_timings:      sample time =      59.60 ms /   100 runs   (    0.60 ms per token,  1677.77 tokens per second)
llama_print_timings: prompt eval time =     482.47 ms /     8 tokens (   60.31 ms per token,    16.58 tokens per second)
llama_print_timings:        eval time =   16688.46 ms /    99 runs   (  168.57 ms per token,     5.93 tokens per second)
llama_print_timings:       total time =   17273.87 ms /   107 tokens

llama_print_timings:        load time =    4373.76 ms
llama_print_timings:      sample time =      61.13 ms /   100 runs   (    0.61 ms per token,  1635.97 tokens per second)
llama_print_timings: prompt eval time =    2181.57 ms /     8 tokens (  272.70 ms per token,     3.67 tokens per second)
llama_print_timings:        eval time =   16500.22 ms /    99 runs   (  166.67 ms per token,     6.00 tokens per second)
llama_print_timings:       total time =   18787.88 ms /   107 tokens

This approach does achieve a higher (12T/s) total speed.
For some reason, I have to copy the model to another copy as the model for the second process, otherwise the performance of the second process will degrade greatly.


--numa isolate：
$ ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 16 --numa isolate -p "I believe the meaning of life is" --no-mmap -b 65535 -c 4096 -np 4096 -ns 65535 -cb
llama_print_timings:        load time =    7619.42 ms
llama_print_timings:      sample time =      59.18 ms /   100 runs   (    0.59 ms per token,  1689.87 tokens per second)
llama_print_timings: prompt eval time =     578.02 ms /     8 tokens (   72.25 ms per token,    13.84 tokens per second)
llama_print_timings:        eval time =   24663.30 ms /    99 runs   (  249.12 ms per token,     4.01 tokens per second)
llama_print_timings:       total time =   25341.84 ms /   107 tokens

without --no-mmap:
$ ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 16 --numa isolate -p "I believe the meaning of life is"
llama_print_timings:        load time =    5821.06 ms
llama_print_timings:      sample time =      59.42 ms /   100 runs   (    0.59 ms per token,  1682.96 tokens per second)
llama_print_timings: prompt eval time =     488.25 ms /     8 tokens (   61.03 ms per token,    16.38 tokens per second)
llama_print_timings:        eval time =   16641.35 ms /    99 runs   (  168.09 ms per token,     5.95 tokens per second)
llama_print_timings:       total time =   17232.02 ms /   107 tokens


$ ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 32 --numa distribute -p "I believe the meaning of life is" --no-mmap -b 65535 -c 4096 -np 4096 -ns 65535 -cb
llama_print_timings:        load time =    7773.76 ms
llama_print_timings:      sample time =      68.89 ms /   100 runs   (    0.69 ms per token,  1451.51 tokens per second)
llama_print_timings: prompt eval time =     427.82 ms /     8 tokens (   53.48 ms per token,    18.70 tokens per second)
llama_print_timings:        eval time =   20671.27 ms /    99 runs   (  208.80 ms per token,     4.79 tokens per second)
llama_print_timings:       total time =   21214.70 ms /   107 tokens
$ numactl -H
available: 2 nodes (0-1)
node 0 cpus: 0 1 2 3 4 5 6 7 8 9 10 11 12 13 14 15 32 33 34 35 36 37 38 39 40 41 42 43 44 45 46 47
node 0 size: 128788 MB
node 0 free: 111567 MB
node 1 cpus: 16 17 18 19 20 21 22 23 24 25 26 27 28 29 30 31 48 49 50 51 52 53 54 55 56 57 58 59 60 61 62 63
node 1 size: 129010 MB
node 1 free: 126397 MB

There seems to be a problem with node allocation, using only node 0's memory. Found the cause, it was caused by-- no-mmap. 
The main branch seems to have the same problem, which may be the cause of the problem you encountered in the first place.


$ ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 32 --numa distribute -p "I believe the meaning of life is"
llama_print_timings:        load time =    4845.21 ms
llama_print_timings:      sample time =      62.89 ms /   100 runs   (    0.63 ms per token,  1590.08 tokens per second)
llama_print_timings: prompt eval time =     404.77 ms /     8 tokens (   50.60 ms per token,    19.76 tokens per second)
llama_print_timings:        eval time =   22165.60 ms /    99 runs   (  223.89 ms per token,     4.47 tokens per second)
llama_print_timings:       total time =   22680.61 ms /   107 tokens

Without clearing the cache, run it again:
llama_print_timings:        load time =     301.42 ms
llama_print_timings:      sample time =      63.00 ms /   100 runs   (    0.63 ms per token,  1587.35 tokens per second)
llama_print_timings: prompt eval time =     364.73 ms /     8 tokens (   45.59 ms per token,    21.93 tokens per second)
llama_print_timings:        eval time =   12217.43 ms /    99 runs   (  123.41 ms per token,     8.10 tokens per second)
llama_print_timings:       total time =   12689.16 ms /   107 tokens


$  numactl --interleave=all ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 30 -p "I believe the meaning of life is" --no-mmap -b 65535 -c 4096 -np 4096 -ns 65535 -cb
llama_print_timings:        load time =    7505.05 ms
llama_print_timings:      sample time =      63.45 ms /   100 runs   (    0.63 ms per token,  1576.09 tokens per second)
llama_print_timings: prompt eval time =     344.12 ms /     8 tokens (   43.02 ms per token,    23.25 tokens per second)
llama_print_timings:        eval time =   14109.30 ms /    99 runs   (  142.52 ms per token,     7.02 tokens per second)
llama_print_timings:       total time =   14561.08 ms /   107 tokens

without --no-mmap:
$ numactl --interleave=all ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 30 -p "I believe the meaning of life is"
llama_print_timings:        load time =    5265.54 ms
llama_print_timings:      sample time =      63.77 ms /   100 runs   (    0.64 ms per token,  1568.26 tokens per second)
llama_print_timings: prompt eval time =     354.31 ms /     8 tokens (   44.29 ms per token,    22.58 tokens per second)
llama_print_timings:        eval time =   13709.93 ms /    99 runs   (  138.48 ms per token,     7.22 tokens per second)
llama_print_timings:       total time =   14173.59 ms /   107 tokens

numactl --interleave=all  Should be equivalent to your settings in bios


--------------
Mainline branch:
$ ./main -m ../../gguf/llama-2-13b-chat.Q4_K_M.gguf -s 12346 -n 100 -t 32 --numa -p "I believe the meaning of life is"
llama_print_timings:        load time =    5010.85 ms
llama_print_timings:      sample time =      62.25 ms /   100 runs   (    0.62 ms per token,  1606.53 tokens per second)
llama_print_timings: prompt eval time =     387.06 ms /     8 tokens (   48.38 ms per token,    20.67 tokens per second)
llama_print_timings:        eval time =   13536.85 ms /    99 runs   (  136.74 ms per token,     7.31 tokens per second)
llama_print_timings:       total time =   14029.71 ms /   107 tokens

Without clearing the cache, run it again:
llama_print_timings:        load time =     318.90 ms
llama_print_timings:      sample time =      62.96 ms /   100 runs   (    0.63 ms per token,  1588.21 tokens per second)
llama_print_timings: prompt eval time =     368.81 ms /     8 tokens (   46.10 ms per token,    21.69 tokens per second)
llama_print_timings:        eval time =   18458.69 ms /    99 runs   (  186.45 ms per token,     5.36 tokens per second)
llama_print_timings:       total time =   18934.58 ms /   107 tokens

Without clearing the cache, run it again:
llama_print_timings:        load time =     343.16 ms
llama_print_timings:      sample time =      63.70 ms /   100 runs   (    0.64 ms per token,  1569.78 tokens per second)
llama_print_timings: prompt eval time =     366.23 ms /     8 tokens (   45.78 ms per token,    21.84 tokens per second)
llama_print_timings:        eval time =   18020.94 ms /    99 runs   (  182.03 ms per token,     5.49 tokens per second)
llama_print_timings:       total time =   18494.47 ms /   107 tokens